added option to keep buildconfigs after builds are complete

+++ openshift/scripts/provision-demo.sh
@@ -32,6 +32,7 @@ function usage() {
     echo "   --project-suffix [suffix] Suffix to be added to demo project names e.g. ci-SUFFIX. If empty, user will be used as suffix"
     echo "   --ephemeral               Deploy demo without persistent storage"
     echo "   --run-verify              Run verify after provisioning"
+    echo "   --keep-builds             Keep the coolstore build configs after builds are complete"
     echo
 }
 
@@ -42,6 +43,7 @@ ARG_EPHEMERAL=false
 ARG_COMMAND=
 ARG_RUN_VERIFY=false
 ARG_DEMO=
+ARG_DELETE_BUILDS=true
 
 while :; do
     case $1 in
@@ -120,6 +122,9 @@ while :; do
             ;;
         --run-verify)
             ARG_RUN_VERIFY=true
+            ;;
+        --keep-builds)
+            ARG_DELETE_BUILDS=false
             ;;
         -h|--help)
             usage
@@ -611,7 +616,9 @@ function promote_images() {
   echo_header "Promoting Images ..."
 
   # remove buildconfigs
-  oc delete bc --all -n ${PRJ_COOLSTORE_PROD[0]}
+  if [ "$ARG_DELETE_BUILDS" = true ] ; then
+    oc delete bc --all -n ${PRJ_COOLSTORE_PROD[0]}
+  fi
 
   for is in coolstore-gw web-ui cart catalog pricing rating review
   do
